fixing select one page
