+++ pages/api/flat/new.js
@@ -2,7 +2,7 @@ import flatController from "../../../controllers/flatController"
 export default async function handler(req, res) {
     // TODO take the title and description from the request body
     const{title, description, location, price, codeNo, image} = req.body
-    console.log(title, description, location, price, codeNo, image)
+/*     console.log(title, description, location, price, codeNo, image) */
     const flat = await flatController.create({
       title, description, location, price, codeNo, image
     })
